smart_cotton_system/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0

# Task 3 AI Bot dependencies
beautifulsoup4==4.14.3
tensorflow==2.20.0
scikit-learn==1.7.2
numpy==2.3.5
pandas==2.3.3
lxml==6.0.2

# ML Quality Control dependencies
joblib==1.4.2
opencv-python==4.11.0.86
xgboost==2.1.3

# Market analysis dependencies
yfinance

# Route optimization dependencies (Task 6)
google-generativeai==0.8.3